+++ changelog.md
@@ -6,6 +6,7 @@
 
 - **Botsonic** plugin (commercial)
 - #745 Page Builder: depend story visibility on user roles.
+- Added option to configure allowed characters for user user names 
 
 ### Improvements